chore: linting should be ay-oh-kay (#21)
Configured a linter to make sure our codebase is styled similarly

- [x] config file for the eslint
- [ ] ~prettier.yml~ 
- [x] stylelint
- [x] github action to run `yarn lint`

We decided that a `prettier.yml` file would cause a whole lot of confusion. Mostly because the prettier.yml file would have to be in sync with the `.eslintrc.js` file. 

We also decided that we should not auto run `yarn lint --fix` because we never want to commit directly to someone else's work.

Fixes #16